[BUGFIX] Updates internals for Ember Canary
We just landed some pretty major changes to the Descriptor system
upstream. We verified that ember-concurrency would still be able to work
first, this breanch is that work. We should likely update it to use
`ember-compatibility-helpers` as well, so you can continue to support
older versions of Ember at the same time.